backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.11.0
pydantic-settings==2.5.2
openai-agents==0.1.0
openai>=1.87.0
psycopg2-binary==2.9.10

